services/payment-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
pydantic==2.7.1
sqlalchemy==2.0.30
asyncpg==0.29.0
redis==5.0.4
structlog==24.1.0
python-dotenv==1.0.1
